LocaleMixin for testing, added to running totals.
